clusterann: make the encoding configurable, move QuantizationParams into format/ (formats 9009df3)

+++ src/main/java/org/opensearch/knn/clusterann/write/block/scalar/OptimizedScalarQuantizedClusterWriter.java
@@ -54,7 +54,7 @@ public final class OptimizedScalarQuantizedClusterWriter implements ClusterWrite
         final int blockSize,
         final int dimension,
         final VectorSimilarityFunction metric,
-        final byte docBits,
+        final int docBits,
         final float[] centroid
     ) {
         this.blockSize = blockSize;
